@@ -1,16 +1,26 @@
 name: dio
-description: A powerful Http client for Dart, which supports Interceptors, FormData, Request Cancellation, File Downloading, Timeout etc.
-version: 4.0.6
-homepage: https://github.com/flutterchina/dio
+
+description: |
+  A powerful HTTP package for Dart/Flutter,
+  which supports Global settings, Interceptors,
+  FormData, Aborting and canceling a request,
+  Files uploading and downloading,
+  Requests timeout, Custom adapters, etc.
+homepage: https://github.com/cfug/dio
+repository: https://github.com/cfug/dio/blob/main/dio
+issue_tracker: https://github.com/cfug/dio/issues
+version: 5.0.3
 
 environment:
-  sdk: '>=2.12.1 <3.0.0'
+  sdk: '>=2.15.0 <3.0.0'
 
 dependencies:
   http_parser: ^4.0.0
+  meta: ^1.5.0
   path: ^1.8.0
 
 dev_dependencies:
-  lints: ^1.0.1
+  lints: any
   test: ^1.5.1
   coverage: ^1.0.3
+  crypto: ^3.0.2
