@@ -1,23 +1,26 @@
 name: dio_example
-description: A powerful Http client for Dart, which supports Interceptors, FormData, Request Cancellation, File Downloading, Timeout etc.
+description: The example snippets for Dio.
 version: 0.0.1
 publish_to: "none"
-homepage: https://github.com/flutterchina/dio
 
 environment:
-  sdk: ">=2.12.1 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
+  cookie_jar:
+  crypto:
   dio:
+  dio_cookie_manager:
+  dio_http2_adapter:
+
+dependency_overrides:
+  dio:
+    path: ../dio
   dio_cookie_manager:
     path: ../plugins/cookie_manager
   dio_http2_adapter:
     path: ../plugins/http2_adapter
 
-dependency_overrides:
-  dio:
-    path: ../dio
-
 dev_dependencies:
-  lints: ^1.0.1
-  test: ^1.5.1
+  lints: any
+  test: ^1.22.1
