@@ -1,21 +1,17 @@
 name: dio_cookie_manager
 description: A cookie manager for Dio, which supports persistent cookies in RAM and file.
-version: 2.0.0
-homepage:  https://github.com/flutterchina/dio/tree/master/plugins/cookie_manager
+version: 2.1.3
+homepage: https://github.com/cfug/dio
+repository: https://github.com/cfug/dio/blob/main/plugins/cookie_manager
+issue_tracker: https://github.com/cfug/dio/issues
 
 environment:
-  sdk: '>=2.12.1 <3.0.0'
+  sdk: '>=2.15.0 <3.0.0'
 
 dependencies:
   cookie_jar: ^3.0.0
-  dio: ^4.0.0
-
-dependency_overrides:
-  dio:
-    path: ../../dio
+  dio: ^5.0.0
 
 dev_dependencies:
-  lints: ^1.0.1
+  lints: any
   test: ^1.16.4
-
-
