@@ -1,4 +1,5 @@
 import 'dart:async';
+
 import 'adapter.dart';
 import 'dio_mixin.dart';
 import 'options.dart';
@@ -6,11 +7,9 @@ import 'headers.dart';
 import 'cancel_token.dart';
 import 'transformer.dart';
 import 'response.dart';
-import 'entry_stub.dart'
-// ignore: uri_does_not_exist
-    if (dart.library.html) 'entry/dio_for_browser.dart'
-// ignore: uri_does_not_exist
-    if (dart.library.io) 'entry/dio_for_native.dart';
+
+import 'dio/dio_for_native.dart'
+    if (dart.library.html) 'dio/dio_for_browser.dart';
 
 /// A powerful Http client for Dart, which supports Interceptors,
 /// Global configuration, FormData, File downloading etc. and Dio is
@@ -20,19 +19,19 @@ import 'entry_stub.dart'
 /// 1. create first , then config it
 ///
 ///   ```dart
-///    var dio = Dio();
+///    final dio = Dio();
 ///    dio.options.baseUrl = "http://www.dtworkroom.com/doris/1/2.0.0/";
-///    dio.options.connectTimeout = 5000; //5s
-///    dio.options.receiveTimeout = 5000;
+///    dio.options.connectTimeout = const Duration(seconds: 5);
+///    dio.options.receiveTimeout = const Duration(seconds: 5);
 ///    dio.options.headers = {HttpHeaders.userAgentHeader: 'dio', 'common-header': 'xx'};
 ///   ```
 /// 2. create and config it:
 ///
 /// ```dart
-///   var dio = Dio(BaseOptions(
+///   final dio = Dio(BaseOptions(
 ///    baseUrl: "http://www.dtworkroom.com/doris/1/2.0.0/",
-///    connectTimeout: 5000,
-///    receiveTimeout: 5000,
+///    connectTimeout: const Duration(seconds: 5),
+///    receiveTimeout: const Duration(seconds: 5),
 ///    headers: {HttpHeaders.userAgentHeader: 'dio', 'common-header': 'xx'},
 ///   ));
 ///  ```
@@ -64,6 +63,7 @@ abstract class Dio {
   /// Handy method to make http GET request, which is a alias of  [dio.fetch(RequestOptions)].
   Future<Response<T>> get<T>(
     String path, {
+    Object? data,
     Map<String, dynamic>? queryParameters,
     Options? options,
     CancelToken? cancelToken,
@@ -73,6 +73,7 @@ abstract class Dio {
   /// Handy method to make http GET request, which is a alias of [dio.fetch(RequestOptions)].
   Future<Response<T>> getUri<T>(
     Uri uri, {
+    Object? data,
     Options? options,
     CancelToken? cancelToken,
     ProgressCallback? onReceiveProgress,
@@ -81,7 +82,7 @@ abstract class Dio {
   /// Handy method to make http POST request, which is a alias of  [dio.fetch(RequestOptions)].
   Future<Response<T>> post<T>(
     String path, {
-    data,
+    Object? data,
     Map<String, dynamic>? queryParameters,
     Options? options,
     CancelToken? cancelToken,
@@ -92,7 +93,7 @@ abstract class Dio {
   /// Handy method to make http POST request, which is a alias of  [dio.fetch(RequestOptions)].
   Future<Response<T>> postUri<T>(
     Uri uri, {
-    data,
+    Object? data,
     Options? options,
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
@@ -102,7 +103,7 @@ abstract class Dio {
   /// Handy method to make http PUT request, which is a alias of  [dio.fetch(RequestOptions)].
   Future<Response<T>> put<T>(
     String path, {
-    data,
+    Object? data,
     Map<String, dynamic>? queryParameters,
     Options? options,
     CancelToken? cancelToken,
@@ -113,7 +114,7 @@ abstract class Dio {
   /// Handy method to make http PUT request, which is a alias of  [dio.fetch(RequestOptions)].
   Future<Response<T>> putUri<T>(
     Uri uri, {
-    data,
+    Object? data,
     Options? options,
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
@@ -123,7 +124,7 @@ abstract class Dio {
   /// Handy method to make http HEAD request, which is a alias of [dio.fetch(RequestOptions)].
   Future<Response<T>> head<T>(
     String path, {
-    data,
+    Object? data,
     Map<String, dynamic>? queryParameters,
     Options? options,
     CancelToken? cancelToken,
@@ -132,7 +133,7 @@ abstract class Dio {
   /// Handy method to make http HEAD request, which is a alias of [dio.fetch(RequestOptions)].
   Future<Response<T>> headUri<T>(
     Uri uri, {
-    data,
+    Object? data,
     Options? options,
     CancelToken? cancelToken,
   });
@@ -140,7 +141,7 @@ abstract class Dio {
   /// Handy method to make http DELETE request, which is a alias of  [dio.fetch(RequestOptions)].
   Future<Response<T>> delete<T>(
     String path, {
-    data,
+    Object? data,
     Map<String, dynamic>? queryParameters,
     Options? options,
     CancelToken? cancelToken,
@@ -149,7 +150,7 @@ abstract class Dio {
   /// Handy method to make http DELETE request, which is a alias of  [dio.fetch(RequestOptions)].
   Future<Response<T>> deleteUri<T>(
     Uri uri, {
-    data,
+    Object? data,
     Options? options,
     CancelToken? cancelToken,
   });
@@ -157,7 +158,7 @@ abstract class Dio {
   /// Handy method to make http PATCH request, which is a alias of  [dio.fetch(RequestOptions)].
   Future<Response<T>> patch<T>(
     String path, {
-    data,
+    Object? data,
     Map<String, dynamic>? queryParameters,
     Options? options,
     CancelToken? cancelToken,
@@ -168,139 +169,102 @@ abstract class Dio {
   /// Handy method to make http PATCH request, which is a alias of  [dio.fetch(RequestOptions)].
   Future<Response<T>> patchUri<T>(
     Uri uri, {
-    data,
+    Object? data,
     Options? options,
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
   });
 
-  /// Lock the current Dio instance.
+  /// {@template dio.Dio.download}
+  /// Download the file and save it in local. The default http method is "GET",
+  /// you can custom it by [Options.method].
   ///
-  /// Dio will enqueue the incoming request tasks instead
-  /// send them directly when [interceptor.requestOptions] is locked.
-  @Deprecated(
-      'Will delete in v5.0. Use `QueuedInterceptor` instead, more detail see'
-      ' https://github.com/flutterchina/dio/issues/1308')
-  void lock();
-
-  /// Unlock the current Dio instance.
+  /// [urlPath] is the file url.
   ///
-  /// Dio instance dequeue the request task。
-  @Deprecated(
-      'Will delete in v5.0. Use `QueuedInterceptor` instead, more detail see'
-      ' https://github.com/flutterchina/dio/issues/1308')
-  void unlock();
-
-  ///Clear the current Dio instance waiting queue.
-  @Deprecated(
-      'Will delete in v5.0. Use `QueuedInterceptor` instead, more detail see'
-      ' https://github.com/flutterchina/dio/issues/1308')
-  void clear();
-
-  ///  Download the file and save it in local. The default http method is "GET",
-  ///  you can custom it by [Options.method].
+  /// [savePath] is the path to save the downloading file later. it can be a String or
+  /// a callback:
+  /// 1. A path with String type, eg "xs.jpg"
+  /// 2. A callback `String Function(Headers headers)`; for example:
+  /// ```dart
+  /// await dio.download(
+  ///   url,
+  ///   (Headers headers) {
+  ///     // Extra info: redirect counts
+  ///     print(headers.value('redirects'));
+  ///     // Extra info: real uri
+  ///     print(headers.value('uri'));
+  ///     // ...
+  ///     return (await getTemporaryDirectory()).path + 'file_name';
+  ///   },
+  /// );
+  /// ```
   ///
-  ///  [urlPath]: The file url.
+  /// [onReceiveProgress] is the callback to listen downloading progress.
+  /// Please refer to [ProgressCallback].
   ///
-  ///  [savePath]: The path to save the downloading file later. it can be a String or
-  ///  a callback:
-  ///  1. A path with String type, eg "xs.jpg"
-  ///  2. A callback `String Function(Headers headers)`; for example:
-  ///  ```dart
-  ///   await dio.download(url,(Headers headers){
-  ///        // Extra info: redirect counts
-  ///        print(headers.value('redirects'));
-  ///        // Extra info: real uri
-  ///        print(headers.value('uri'));
-  ///      ...
-  ///      return "...";
-  ///    });
-  ///  ```
+  /// [deleteOnError] whether delete the file when error occurs.
+  /// The default value is [true].
   ///
-  ///  [onReceiveProgress]: The callback to listen downloading progress.
-  ///  please refer to [ProgressCallback].
+  /// [lengthHeader] : The real size of original file (not compressed).
+  /// When file is compressed:
+  /// 1. If this value is 'content-length', the `total` argument of [onReceiveProgress] will be -1
+  /// 2. If this value is not 'content-length', maybe a custom header indicates the original
+  /// file size , the `total` argument of [onReceiveProgress] will be this header value.
   ///
-  /// [deleteOnError] Whether delete the file when error occurs. The default value is [true].
+  /// You can also disable the compression by specifying
+  /// the 'accept-encoding' header value as '*' to assure the value of
+  /// `total` argument of [onReceiveProgress] is not -1. for example:
   ///
-  ///  [lengthHeader] : The real size of original file (not compressed).
-  ///  When file is compressed:
-  ///  1. If this value is 'content-length', the `total` argument of `onProgress` will be -1
-  ///  2. If this value is not 'content-length', maybe a custom header indicates the original
-  ///  file size , the `total` argument of `onProgress` will be this header value.
-  ///
-  ///  you can also disable the compression by specifying the 'accept-encoding' header value as '*'
-  ///  to assure the value of `total` argument of `onProgress` is not -1. for example:
-  ///
-  ///     await dio.download(url, "./example/flutter.svg",
-  ///     options: Options(headers: {HttpHeaders.acceptEncodingHeader: "*"}),  // disable gzip
-  ///     onProgress: (received, total) {
-  ///       if (total != -1) {
-  ///        print((received / total * 100).toStringAsFixed(0) + "%");
-  ///       }
-  ///     });
-
+  /// ```dart
+  /// await dio.download(
+  ///   url,
+  ///   (await getTemporaryDirectory()).path + 'flutter.svg',
+  ///   options: Options(
+  ///     headers: {HttpHeaders.acceptEncodingHeader: "*"}, // Disable gzip
+  ///   ),
+  ///   onReceiveProgress: (received, total) {
+  ///     if (total != -1) {
+  ///       print((received / total * 100).toStringAsFixed(0) + "%");
+  ///     }
+  ///   },
+  /// );
+  /// ```
+  /// {@endtemplate}
   Future<Response> download(
     String urlPath,
-    savePath, {
+    dynamic savePath, {
     ProgressCallback? onReceiveProgress,
     Map<String, dynamic>? queryParameters,
     CancelToken? cancelToken,
     bool deleteOnError = true,
     String lengthHeader = Headers.contentLengthHeader,
-    data,
+    Object? data,
     Options? options,
   });
 
-  ///  Download the file and save it in local. The default http method is "GET",
-  ///  you can custom it by [Options.method].
-  ///
-  ///  [uri]: The file url.
-  ///
-  ///  [savePath]: The path to save the downloading file later. it can be a String or
-  ///  a callback:
-  ///  1. A path with String type, eg "xs.jpg"
-  ///  2. A callback `String Function(Headers)`; for example:
-  ///  ```dart
-  ///   await dio.downloadUri(uri,(Headers headers){
-  ///        // Extra info: redirect counts
-  ///        print(headers.value('redirects'));
-  ///        // Extra info: real uri
-  ///        print(headers.value('uri'));
-  ///       ...
-  ///       return "...";
-  ///    });
-  ///  ```
-  ///
-  ///  [onReceiveProgress]: The callback to listen downloading progress.
-  ///  please refer to [ProgressCallback].
-  ///
-  ///  [lengthHeader] : The real size of original file (not compressed).
-  ///  When file is compressed:
-  ///  1. If this value is 'content-length', the `total` argument of `onProgress` will be -1
-  ///  2. If this value is not 'content-length', maybe a custom header indicates the original
-  ///  file size , the `total` argument of `onProgress` will be this header value.
-  ///
-  ///  you can also disable the compression by specifying the 'accept-encoding' header value as '*'
-  ///  to assure the value of `total` argument of `onProgress` is not -1. for example:
-  ///
-  ///     await dio.downloadUri(uri, "./example/flutter.svg",
-  ///     options: Options(headers: {HttpHeaders.acceptEncodingHeader: "*"}),  // disable gzip
-  ///     onProgress: (received, total) {
-  ///       if (total != -1) {
-  ///        print((received / total * 100).toStringAsFixed(0) + "%");
-  ///       }
-  ///     });
+  /// {@macro dio.Dio.download}
   Future<Response> downloadUri(
     Uri uri,
-    savePath, {
+    dynamic savePath, {
     ProgressCallback? onReceiveProgress,
     CancelToken? cancelToken,
     bool deleteOnError = true,
     String lengthHeader = Headers.contentLengthHeader,
-    data,
+    Object? data,
     Options? options,
-  });
+  }) {
+    return download(
+      uri.toString(),
+      savePath,
+      onReceiveProgress: onReceiveProgress,
+      lengthHeader: lengthHeader,
+      deleteOnError: deleteOnError,
+      cancelToken: cancelToken,
+      data: data,
+      options: options,
+    );
+  }
 
   /// Make http request with options.
   ///
@@ -309,7 +273,7 @@ abstract class Dio {
   /// [options] The request options.
   Future<Response<T>> request<T>(
     String path, {
-    data,
+    Object? data,
     Map<String, dynamic>? queryParameters,
     CancelToken? cancelToken,
     Options? options,
@@ -324,7 +288,7 @@ abstract class Dio {
   /// [options] The request options.
   Future<Response<T>> requestUri<T>(
     Uri uri, {
-    data,
+    Object? data,
     CancelToken? cancelToken,
     Options? options,
     ProgressCallback? onSendProgress,
