@@ -7,29 +7,22 @@ import 'redirect_record.dart';
 class Response<T> {
   Response({
     this.data,
-    Headers? headers,
     required this.requestOptions,
-    this.isRedirect,
     this.statusCode,
     this.statusMessage,
-    List<RedirectRecord>? redirects,
-    Map<String, dynamic>? extra,
-  }) {
-    this.headers = headers ?? Headers();
-    this.extra = extra ?? {};
-    this.redirects = redirects ?? [];
-  }
+    this.isRedirect = false,
+    this.redirects = const [],
+    this.extra = const {},
+    Headers? headers,
+  }) : headers = headers ?? Headers();
 
   /// Response body. may have been transformed, please refer to [ResponseType].
   T? data;
 
-  /// Response headers.
-  late Headers headers;
-
   /// The corresponding request info.
-  late RequestOptions requestOptions;
+  RequestOptions requestOptions;
 
-  /// Http status code.
+  /// HTTP status code.
   int? statusCode;
 
   /// Returns the reason phrase associated with the status code.
@@ -37,30 +30,33 @@ class Response<T> {
   /// to. Setting the reason phrase after writing to the body.
   String? statusMessage;
 
-  /// Custom field that you can retrieve it later in `then`.
-  late Map<String, dynamic> extra;
-
-  /// Returns the series of redirects this connection has been through. The
-  /// list will be empty if no redirects were followed. [redirects] will be
-  /// updated both in the case of an automatic and a manual redirect.
-  ///
-  /// ** Attention **: Whether this field is available depends on whether the
-  /// implementation of the adapter supports it or not.
-  late List<RedirectRecord> redirects;
-
   /// Whether this response is a redirect.
   /// ** Attention **: Whether this field is available depends on whether the
   /// implementation of the adapter supports it or not.
-  bool? isRedirect;
+  bool isRedirect;
 
-  /// Return the final real request uri (maybe redirect).
+  /// The series of redirects this connection has been through. The list will be
+  /// empty if no redirects were followed. [redirects] will be updated both
+  /// in the case of an automatic and a manual redirect.
   ///
   /// ** Attention **: Whether this field is available depends on whether the
   /// implementation of the adapter supports it or not.
-  Uri get realUri =>
-      (redirects.isNotEmpty) ? redirects.last.location : requestOptions.uri;
+  List<RedirectRecord> redirects;
 
-  /// We are more concerned about `data` field.
+  /// Custom fields that are constructed in the [RequestOptions].
+  Map<String, dynamic> extra;
+
+  /// Response headers.
+  Headers headers;
+
+  /// Return the final real request URI (may be redirected).
+  ///
+  /// Note: Whether the field is available depends on whether the adapter
+  /// supports or not.
+  Uri get realUri =>
+      redirects.isNotEmpty ? redirects.last.location : requestOptions.uri;
+
+  /// We are more concerned about [data] field.
   @override
   String toString() {
     if (data is Map) {
