@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:test/test.dart';
 
 void main() {
-  test('#assureResponse', () {
+  test('assureResponse', () {
     final untypedResponse = Response<dynamic>(
       requestOptions: RequestOptions(path: ''),
       data: null,
