@@ -1,21 +1,19 @@
-// Copyright (c) 2014, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
 @TestOn('vm')
 import 'dart:io';
 
 import 'package:dio/dio.dart';
+import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
+import 'mock/adapters.dart';
 import 'utils.dart';
 
 void main() {
   setUp(startServer);
   tearDown(stopServer);
-  test('#test download1', () async {
+  test('download1', () async {
     const savePath = 'test/_download_test.md';
-    var dio = Dio();
+    final dio = Dio();
     dio.options.baseUrl = serverUrl.toString();
     await dio.download(
       '/download', savePath, // disable gzip
@@ -24,73 +22,96 @@ void main() {
       },
     );
 
-    var f = File(savePath);
+    final f = File(savePath);
     expect(f.readAsStringSync(), equals('I am a text file'));
     f.deleteSync(recursive: false);
   });
 
-  test('#test download2', () async {
+  test('download2', () async {
     const savePath = 'test/_download_test.md';
-    var dio = Dio();
+    final dio = Dio();
     dio.options.baseUrl = serverUrl.toString();
     await dio.downloadUri(
       serverUrl.replace(path: '/download'),
       (header) => savePath, // disable gzip
     );
 
-    var f = File(savePath);
+    final f = File(savePath);
     expect(f.readAsStringSync(), equals('I am a text file'));
     f.deleteSync(recursive: false);
   });
 
-  test('#test download error', () async {
+  test('download error', () async {
     const savePath = 'test/_download_test.md';
-    var dio = Dio();
+    final dio = Dio();
     dio.options.baseUrl = serverUrl.toString();
-    var r = await dio
+    Response response = await dio
         .download('/error', savePath)
         .catchError((e) => (e as DioError).response!);
-    assert(r.data == 'error');
-    r = await dio
+    expect(response.data, 'error');
+    response = await dio
         .download(
           '/error',
           savePath,
           options: Options(receiveDataWhenStatusError: false),
         )
         .catchError((e) => (e as DioError).response!);
-    assert(r.data == null);
+    expect(response.data, null);
   });
 
-  test('#test download timeout', () async {
+  test('download timeout', () async {
     const savePath = 'test/_download_test.md';
-    var dio = Dio(BaseOptions(
-      receiveTimeout: 1,
-      baseUrl: serverUrl.toString(),
-    ));
+    final dio = Dio(
+      BaseOptions(
+        receiveTimeout: Duration(milliseconds: 1),
+        baseUrl: serverUrl.toString(),
+      ),
+    );
     expect(
-        dio
-            .download('/download', savePath)
-            .catchError((e) => throw (e as DioError).type),
-        throwsA(DioErrorType.receiveTimeout));
+      dio
+          .download('/download', savePath)
+          .catchError((e) => throw (e as DioError).type),
+      throwsA(DioErrorType.receiveTimeout),
+    );
     //print(r);
   });
 
-  test('#test download cancellation', () async {
+  test('download cancellation', () async {
     const savePath = 'test/_download_test.md';
-    var cancelToken = CancelToken();
+    final cancelToken = CancelToken();
     Future.delayed(Duration(milliseconds: 100), () {
       cancelToken.cancel();
     });
     expect(
       Dio()
           .download(
-            serverUrl.toString() + '/download',
+            '$serverUrl/download',
             savePath,
             cancelToken: cancelToken,
           )
           .catchError((e) => throw (e as DioError).type),
       throwsA(DioErrorType.cancel),
     );
-    //print(r);
+  });
+
+  test('`savePath` types', () async {
+    final testPath = p.join(Directory.systemTemp.path, 'dio', 'testPath');
+
+    final dio = Dio()
+      ..options.baseUrl = EchoAdapter.mockBase
+      ..httpClientAdapter = EchoAdapter();
+
+    await expectLater(
+      dio.download('/test', testPath),
+      completes,
+    );
+    await expectLater(
+      dio.download('/test', (headers) => testPath),
+      completes,
+    );
+    await expectLater(
+      dio.download('/test', (headers) async => testPath),
+      completes,
+    );
   });
 }
