@@ -1,7 +1,3 @@
-// Copyright (c) 2014, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
@@ -22,8 +18,8 @@ Uri get serverUrl => Uri.parse('http://localhost:${_server?.port}');
 Future<void> startServer() async {
   _server = (await HttpServer.bind('localhost', 0))
     ..listen((request) async {
-      var path = request.uri.path;
-      var response = request.response;
+      final path = request.uri.path;
+      final response = request.response;
 
       if (path == '/error') {
         const content = 'error';
@@ -36,7 +32,7 @@ Future<void> startServer() async {
       }
 
       if (path == '/loop') {
-        var n = int.parse(request.uri.query);
+        final n = int.parse(request.uri.query);
         response
           ..statusCode = 302
           ..headers
@@ -74,6 +70,20 @@ Future<void> startServer() async {
         return;
       }
 
+      if (path == '/multi-value-header') {
+        response.headers.contentType = ContentType('application', 'json');
+        response.headers.add(
+          'x-multi-value-request-header-echo',
+          request.headers.value('x-multi-value-request-header').toString(),
+        );
+        response
+          ..statusCode = 200
+          ..contentLength = -1
+          ..write('');
+        response.close();
+        return;
+      }
+
       if (path == '/download') {
         const content = 'I am a text file';
         response.headers.set('content-encoding', 'plain');
@@ -88,9 +98,9 @@ Future<void> startServer() async {
         return;
       }
 
-      var requestBodyBytes = await ByteStream(request).toBytes();
-      var encodingName = request.uri.queryParameters['response-encoding'];
-      var outputEncoding = encodingName == null
+      final requestBodyBytes = await ByteStream(request).toBytes();
+      final encodingName = request.uri.queryParameters['response-encoding'];
+      final outputEncoding = encodingName == null
           ? ascii
           : requiredEncodingForCharset(encodingName);
 
@@ -101,15 +111,14 @@ Future<void> startServer() async {
       dynamic requestBody;
       if (requestBodyBytes.isEmpty) {
         requestBody = null;
-      } else if (request.headers.contentType?.charset != null) {
-        var encoding =
-            requiredEncodingForCharset(request.headers.contentType!.charset!);
-        requestBody = encoding.decode(requestBodyBytes);
       } else {
-        requestBody = requestBodyBytes;
+        final encoding = requiredEncodingForCharset(
+          request.headers.contentType?.charset ?? 'utf-8',
+        );
+        requestBody = encoding.decode(requestBodyBytes);
       }
 
-      var content = <String, dynamic>{
+      final content = <String, dynamic>{
         'method': request.method,
         'path': request.uri.path,
         'query': request.uri.query,
@@ -124,7 +133,7 @@ Future<void> startServer() async {
         content['headers'][name] = values;
       });
 
-      var body = json.encode(content);
+      final body = json.encode(content);
       response
         ..contentLength = body.length
         ..write(body);
@@ -155,13 +164,16 @@ class ByteStream extends StreamView<List<int>> {
 
   /// Collects the data of this stream in a [Uint8List].
   Future<Uint8List> toBytes() {
-    var completer = Completer<Uint8List>();
-    var sink = ByteConversionSink.withCallback(
-        (bytes) => completer.complete(Uint8List.fromList(bytes)));
-    listen(sink.add,
-        onError: completer.completeError,
-        onDone: sink.close,
-        cancelOnError: true);
+    final completer = Completer<Uint8List>();
+    final sink = ByteConversionSink.withCallback(
+      (bytes) => completer.complete(Uint8List.fromList(bytes)),
+    );
+    listen(
+      sink.add,
+      onError: completer.completeError,
+      onDone: sink.close,
+      cancelOnError: true,
+    );
     return completer.future;
   }
 
