@@ -7,7 +7,7 @@ class CacheInterceptor extends Interceptor {
 
   @override
   void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
-    var response = _cache[options.uri];
+    final response = _cache[options.uri];
     if (options.extra['refresh'] == true) {
       print('${options.uri}: force refresh, ignore cache! \n');
       return handler.next(options);
@@ -32,8 +32,8 @@ class CacheInterceptor extends Interceptor {
 }
 
 void main() async {
-  var dio = Dio();
-  dio.options.baseUrl = 'https://baidu.com';
+  final dio = Dio();
+  dio.options.baseUrl = 'https://pub.dev';
   dio.interceptors
     ..add(CacheInterceptor())
     ..add(LogInterceptor(requestHeader: false, responseHeader: false));
