@@ -1,25 +1,20 @@
 import 'dart:async';
 import 'dart:io';
+
 import 'package:dio/dio.dart';
 
 // In this example we download a image and listen the downloading progress.
 void main() async {
-  var dio = Dio();
+  final dio = Dio();
   dio.interceptors.add(LogInterceptor());
-  // This is big file(about 200M)
-  //   var url = "http://download.dcloud.net.cn/HBuilder.9.0.2.macosx_64.dmg";
-
-  var url =
-      'https://cdn.jsdelivr.net/gh/flutterchina/flutter-in-action@1.0/docs/imgs/book.jpg';
-
-  // var url = "https://www.baidu.com/img/bdlogo.gif";
-  await download1(dio, url, './example/book.jpg');
-  await download1(dio, url, (Headers headers) => './example/book1.jpg');
-  await download2(dio, url, './example/book2.jpg');
+  final url = 'https://pub.dev/static/hash-rhob5slb/img/pub-dev-logo.svg';
+  await download1(dio, url, './example/pub-dev-logo.svg');
+  await download1(dio, url, (headers) => './example/pub-dev-logo-1.svg');
+  await download1(dio, url, (headers) async => './example/pub-dev-logo-2.svg');
 }
 
 Future download1(Dio dio, String url, savePath) async {
-  var cancelToken = CancelToken();
+  final cancelToken = CancelToken();
   try {
     await dio.download(
       url,
@@ -35,19 +30,19 @@ Future download1(Dio dio, String url, savePath) async {
 //Another way to downloading small file
 Future download2(Dio dio, String url, String savePath) async {
   try {
-    var response = await dio.get(
+    final response = await dio.get(
       url,
       onReceiveProgress: showDownloadProgress,
       //Received data with List<int>
       options: Options(
         responseType: ResponseType.bytes,
         followRedirects: false,
-        receiveTimeout: 0,
+        receiveTimeout: Duration.zero,
       ),
     );
     print(response.headers);
-    var file = File(savePath);
-    var raf = file.openSync(mode: FileMode.write);
+    final file = File(savePath);
+    final raf = file.openSync(mode: FileMode.write);
     // response.data is List<int> type
     raf.writeFromSync(response.data);
     await raf.close();
