@@ -1,10 +1,10 @@
 import 'package:dio/dio.dart';
-import 'package:dio/native_imp.dart';
+import 'package:dio/io.dart';
 
 class HttpService extends DioForNative {
   HttpService([BaseOptions? baseOptions]) : super(baseOptions) {
     options
-      ..baseUrl = 'http://httpbin.org/'
+      ..baseUrl = 'https://httpbin.org/'
       ..contentType = Headers.jsonContentType;
   }
 
@@ -14,7 +14,7 @@ class HttpService extends DioForNative {
 }
 
 void main() async {
-  var httpService = HttpService();
-  var res = await httpService.echo('hello server!');
+  final httpService = HttpService();
+  final res = await httpService.echo('hello server!');
   print(res);
 }
