@@ -2,17 +2,17 @@ import 'package:dio/dio.dart';
 import 'package:dio_http2_adapter/dio_http2_adapter.dart';
 
 void main() async {
-  var dio = Dio()
-    ..options.baseUrl = 'https://google.com'
+  final dio = Dio()
+    ..options.baseUrl = 'https://pub.dev'
     ..interceptors.add(LogInterceptor())
     ..httpClientAdapter = Http2Adapter(
-      ConnectionManager(idleTimeout: 10000),
+      ConnectionManager(idleTimeout: Duration(seconds: 10)),
     );
 
   Response<String> response;
   response = await dio.get('/?xx=6');
-  response.redirects.forEach((e) {
+  for (final e in response.redirects) {
     print('redirect: ${e.statusCode} ${e.location}');
-  });
+  }
   print(response.data);
 }
