@@ -1,28 +1,31 @@
 import 'dart:io';
+
 import 'package:dio/dio.dart';
 
 void main() async {
-  var dio = Dio(BaseOptions(
-    baseUrl: 'http://httpbin.org/',
-    connectTimeout: 5000,
-    receiveTimeout: 100000,
-    // 5s
-    headers: {
-      HttpHeaders.userAgentHeader: 'dio',
-      'api': '1.0.0',
-    },
-    contentType: Headers.jsonContentType,
-    // Transform the response data to a String encoded with UTF8.
-    // The default value is [ResponseType.JSON].
-    responseType: ResponseType.plain,
-  ));
+  final dio = Dio(
+    BaseOptions(
+      baseUrl: 'https://httpbin.org/',
+      connectTimeout: Duration(seconds: 5),
+      receiveTimeout: Duration(seconds: 10),
+      // 5s
+      headers: {
+        HttpHeaders.userAgentHeader: 'dio',
+        'api': '1.0.0',
+      },
+      contentType: Headers.jsonContentType,
+      // Transform the response data to a String encoded with UTF8.
+      // The default value is [ResponseType.JSON].
+      responseType: ResponseType.plain,
+    ),
+  );
 
   Response response;
 
   response = await dio.get('/get');
   print(response.data);
 
-  var responseMap = await dio.get(
+  final responseMap = await dio.get(
     '/get',
     // Transform response data to Json Map
     options: Options(responseType: ResponseType.json),
