@@ -1,75 +1,92 @@
 import 'dart:async';
+import 'dart:convert';
 
 import 'package:dio/dio.dart';
 
 void main() async {
-  var dio = Dio();
+  final dio = Dio();
   //  dio instance to request token
-  var tokenDio = Dio();
+  final tokenDio = Dio();
   String? csrfToken;
-  dio.options.baseUrl = 'http://www.dtworkroom.com/doris/1/2.0.0/';
+  dio.options.baseUrl = 'https://seunghwanlytest.mocklab.io/';
   tokenDio.options = dio.options;
-  dio.interceptors.add(QueuedInterceptorsWrapper(
-    onRequest: (options, handler) {
-      print('send request：path:${options.path}，baseURL:${options.baseUrl}');
-      if (csrfToken == null) {
-        print('no token，request token firstly...');
-        tokenDio.get('/token').then((d) {
-          options.headers['csrfToken'] = csrfToken = d.data['data']['token'];
-          print('request token succeed, value: ' + d.data['data']['token']);
-          print(
-              'continue to perform request：path:${options.path}，baseURL:${options.path}');
-          handler.next(options);
-        }).catchError((error, stackTrace) {
-          handler.reject(error, true);
-        });
-      } else {
+  dio.interceptors.add(
+    QueuedInterceptorsWrapper(
+      onRequest: (options, handler) async {
+        print('send request：path:${options.path}，baseURL:${options.baseUrl}');
+
+        if (csrfToken == null) {
+          print('no token，request token firstly...');
+
+          final result = await tokenDio.get('/token');
+
+          if (result.statusCode != null && result.statusCode! ~/ 100 == 2) {
+            /// assume `token` is in response body
+            final body = jsonDecode(result.data) as Map<String, dynamic>?;
+
+            if (body != null && body.containsKey('data')) {
+              options.headers['csrfToken'] = csrfToken = body['data']['token'];
+              print('request token succeed, value: $csrfToken');
+              print(
+                'continue to perform request：path:${options.path}，baseURL:${options.path}',
+              );
+              return handler.next(options);
+            }
+          }
+
+          return handler.reject(
+            DioError(requestOptions: result.requestOptions),
+            true,
+          );
+        }
+
         options.headers['csrfToken'] = csrfToken;
         return handler.next(options);
-      }
-    },
-    onError: (error, handler) {
-      //print(error);
-      // Assume 401 stands for token expired
-      if (error.response?.statusCode == 401) {
-        var options = error.response!.requestOptions;
-        // If the token has been updated, repeat directly.
-        if (csrfToken != options.headers['csrfToken']) {
-          options.headers['csrfToken'] = csrfToken;
-          //repeat
-          dio.fetch(options).then(
-            (r) => handler.resolve(r),
-            onError: (e) {
-              handler.reject(e);
-            },
-          );
-          return;
-        }
-        tokenDio.get('/token').then((d) {
-          //update csrfToken
-          options.headers['csrfToken'] = csrfToken = d.data['data']['token'];
-        }).then((e) {
-          //repeat
-          dio.fetch(options).then(
-            (r) => handler.resolve(r),
-            onError: (e) {
-              handler.reject(e);
-            },
-          );
-        });
-        return;
-      }
-      return handler.next(error);
-    },
-  ));
+      },
+      onError: (error, handler) async {
+        /// Assume 401 stands for token expired
+        if (error.response?.statusCode == 401) {
+          print('the token has expired, need to receive new token');
+          final options = error.response!.requestOptions;
 
-  FutureOr<void> _onResult(d) {
+          /// assume receiving the token has no errors
+          /// to check `null-safety` and error handling
+          /// please check inside the [onRequest] closure
+          final tokenResult = await tokenDio.get('/token');
+
+          /// update [csrfToken]
+          /// assume `token` is in response body
+          final body = jsonDecode(tokenResult.data) as Map<String, dynamic>?;
+          options.headers['csrfToken'] = csrfToken = body!['data']['token'];
+
+          if (options.headers['csrfToken'] != null) {
+            print('the token has been updated');
+
+            /// since the api has no state, force to pass the 401 error
+            /// by adding query parameter
+            final originResult = await dio.fetch(options..path += '&pass=true');
+            if (originResult.statusCode != null &&
+                originResult.statusCode! ~/ 100 == 2) {
+              return handler.resolve(originResult);
+            }
+          }
+          print('the token has not been updated');
+          return handler.reject(
+            DioError(requestOptions: options),
+          );
+        }
+        return handler.next(error);
+      },
+    ),
+  );
+
+  FutureOr<void> onResult(d) {
     print('request ok!');
   }
 
-  await Future.wait([
-    dio.get('/test?tag=1').then(_onResult),
-    dio.get('/test?tag=2').then(_onResult),
-    dio.get('/test?tag=3').then(_onResult)
-  ]);
+  /// assume `/test?tag=2` path occurs the authorization error (401)
+  /// and token to be updated
+  await dio.get('/test?tag=1').then(onResult);
+  await dio.get('/test?tag=2').then(onResult);
+  await dio.get('/test?tag=3').then(onResult);
 }
