@@ -1,35 +1,39 @@
 import 'package:dio/dio.dart';
 
 void main() async {
-  var dio = Dio();
-  dio.options.baseUrl = 'http://httpbin.org/';
-  dio.options.connectTimeout = 5000;
-  dio.interceptors.add(InterceptorsWrapper(onRequest: (options, handler) {
-    switch (options.path) {
-      case '/fakepath1':
-        return handler.resolve(
-          Response(
-            requestOptions: options,
-            data: 'fake data',
-          ),
-        );
-      case '/fakepath2':
-        dio
-            .get('/get')
-            .then(handler.resolve)
-            .catchError((e) => handler.reject(e));
-        break;
-      case '/fakepath3':
-        return handler.reject(
-          DioError(
-            requestOptions: options,
-            error: 'test error',
-          ),
-        );
-      default:
-        return handler.next(options); //continue
-    }
-  }));
+  final dio = Dio();
+  dio.options.baseUrl = 'https://httpbin.org/';
+  dio.options.connectTimeout = Duration(seconds: 5);
+  dio.interceptors.add(
+    InterceptorsWrapper(
+      onRequest: (options, handler) {
+        switch (options.path) {
+          case '/fakepath1':
+            return handler.resolve(
+              Response(
+                requestOptions: options,
+                data: 'fake data',
+              ),
+            );
+          case '/fakepath2':
+            dio
+                .get('/get')
+                .then(handler.resolve)
+                .catchError((e) => handler.reject(e));
+            break;
+          case '/fakepath3':
+            return handler.reject(
+              DioError(
+                requestOptions: options,
+                error: 'test error',
+              ),
+            );
+          default:
+            return handler.next(options); //continue
+        }
+      },
+    ),
+  );
   Response response;
   response = await dio.get('/fakepath1');
   assert(response.data == 'fake data');
@@ -38,14 +42,13 @@ void main() async {
   try {
     response = await dio.get('/fakepath3');
   } on DioError catch (e) {
-    print(1);
     assert(e.message == 'test error');
     assert(e.response == null);
   }
   response = await dio.get('/get');
   assert(response.data['headers'] is Map);
   try {
-    await dio.get('xsddddd');
+    await dio.get('/status/404');
   } on DioError catch (e) {
     assert(e.response!.statusCode == 404);
   }
