@@ -1,12 +1,13 @@
 import 'dart:async';
+
 import 'package:dio/dio.dart';
 
-/// If the request data is a `List` type, the [DefaultTransformer] will send data
+/// If the request data is a `List` type, the [BackgroundTransformer] will send data
 /// by calling its `toString()` method. However, normally the List object is
 /// not expected for request data( mostly need Map ). So we provide a custom
 /// [Transformer] that will throw error when request data is a `List` type.
 
-class MyTransformer extends DefaultTransformer {
+class MyTransformer extends BackgroundTransformer {
   @override
   Future<String> transformRequest(RequestOptions options) async {
     if (options.data is List<String>) {
@@ -33,11 +34,11 @@ class MyTransformer extends DefaultTransformer {
 }
 
 void main() async {
-  var dio = Dio();
+  final dio = Dio();
   // Use custom Transformer
   dio.transformer = MyTransformer();
 
-  var response = await dio.get('https://www.baidu.com');
+  final response = await dio.get('https://www.baidu.com');
   print(response.requestOptions.extra['self']);
 
   try {
