@@ -1,3 +1,7 @@
 import 'package:dio/dio.dart';
 
-var dio = Dio();
+final dio = Dio(
+  BaseOptions(
+    connectTimeout: Duration(seconds: 3),
+  ),
+);
