@@ -5,15 +5,15 @@ import 'package:test/test.dart';
 
 void main() {
   test('cookie-jar', () async {
-    var dio = Dio();
-    var cookieJar = CookieJar();
+    final dio = Dio();
+    final cookieJar = CookieJar();
     dio.interceptors
       ..add(CookieManager(cookieJar))
       ..add(LogInterceptor());
-    await dio.get('https://google.com/');
+    await dio.get('https://pub.dev/');
     // Print cookies
-    print(cookieJar.loadForRequest(Uri.parse('https://google.com/')));
+    print(cookieJar.loadForRequest(Uri.parse('https://pub.dev/')));
     // second request with the cookie
-    await dio.get('https://google.com/');
+    await dio.get('https://pub.dev/');
   });
 }
