@@ -2,14 +2,12 @@ import 'package:dio/dio.dart';
 import 'package:dio_http2_adapter/dio_http2_adapter.dart';
 
 void main() async {
-  var dio = Dio()
-    ..options.baseUrl = 'https://google.com'
+  final dio = Dio()
+    ..options.baseUrl = 'https://pub.dev'
     ..interceptors.add(LogInterceptor())
     ..httpClientAdapter = Http2Adapter(
       ConnectionManager(
-        idleTimeout: 10000,
-        // Ignore bad certificate
-        onClientCreate: (_, config) => config.onBadCertificate = (_) => true,
+        idleTimeout: Duration(seconds: 10),
       ),
     );
 
